+++ R/ir_bc.R
@@ -280,8 +280,8 @@ ir_bc_rubberband <- function(x,
   x %>%
     dplyr::mutate(
       spectra =
-        purrr::map(seq_along(spectra), function(i) {
-          res <- spectra[[i]]
+        purrr::map(seq_along(.data$spectra), function(i) {
+          res <- .data$spectra[[i]]
           y <- x_flat[, i + 1L, drop = TRUE]
           res$y <- y[! is.na(y)]
           res